requirements.txt
streamlit
openai
python-dotenv
langchain
faiss-cpu
PyMuPDF
unstructured
pypdf
playwright
beautifulsoup4
requests
python-pptx
pytesseract
Pillow
pandas
openpyxl
fpdf
plotly
SpeechRecognition